Small reorg in prep for Templates
