+++ cpts/_main.php
@@ -5,9 +5,9 @@
  */
 
 // Include the base files
-require_once 'actors.php';
-require_once 'characters.php';
-require_once 'shows.php';
+require_once 'actors/_main.php';
+require_once 'characters/_main.php';
+require_once 'shows/_main.php';
 require_once 'related-posts.php';
 
 // phpcs:disable
