+++ cpts/shows/_main.php
@@ -487,8 +487,8 @@ SQL;
 }
 
 // Include Sub Files
-require_once 'shows/calculations.php';
-require_once 'shows/cmb2-metaboxes.php';
-require_once 'shows/shows-like-this.php';
+require_once 'calculations.php';
+require_once 'cmb2-metaboxes.php';
+require_once 'shows-like-this.php';
 
 new LWTV_CPT_Shows();
